refactor(gateway): relocate operation result types into core/operations
Extracts ServiceCallResult, the ActionGoalStatus enum + free
action_status_to_string helper, the four Action*Result structs, and
ActionGoalInfo into a dedicated header under core/operations/. The types
were already neutral C++ but lived behind the rclcpp-pulling
operation_manager.hpp include path, which prevented the neutral build
layer from consuming them. The OperationManager header forwards to the
new location; the helper definition moves alongside into a small core
translation unit so the symbol is provided by gateway_core.
